+++ microservices/admin/public/index.html
@@ -30,6 +30,6 @@
       Response: {{ resp.product }}<br/>
       <span v-show="resp.errorCode != 0">Error: {{ resp.errorMsg }}({{ resp.errorCode }})</span>
     </div>
-    <script src="./main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
