+++ microservices/product/Makefile
@@ -1,14 +1,13 @@
 COMMIT_HASH=`git rev-parse --short HEAD`
 
 proto:
-	protoc ./protobuf/product.proto -I ./protobuf --go_out=plugins=grpc:./protobuf --js_out=import_style=commonjs:./client --grpc-web_out=import_style=commonjs,mode=grpcwebtext:./client
+	protoc ./protobuf/product.proto -I ./protobuf --go_out=plugins=grpc:./protobuf --js_out=import_style=commonjs:../admin/src/js/protobuf --grpc-web_out=import_style=commonjs,mode=grpcwebtext:../admin/src/js/protobuf
 
-client:
-	docker build -t k8stestbed/client:$(COMMIT_HASH) ./client
-	docker push k8stestbed/client:$(COMMIT_HASH)
-
-server:
+build:
 	docker build -t k8stestbed/product:$(COMMIT_HASH) .
+
+push: build
 	docker push k8stestbed/product:$(COMMIT_HASH)
 
-.PHONY: proto client server
+
+.PHONY: proto build push
